Create protected routes
- Create RequiredAuth Wrapper to surround routes that need to be authenticated
- Created button for sign out on homepage
- Moved login into its own component so that if a route is not authenticated then a redirect to login will occur
